like comment: backend, frontend

+++ backend/src/main/java/com/example/backend/entity/Comment.java
@@ -15,7 +15,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private Long commentId;
+    private Integer  commentId;
 
     @Column(nullable = false)
     private String commentDescription;
